requirements.txt
# 道路裂缝检测系统依赖包

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# Web界面
gradio>=3.50.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0

# 其他工具
tqdm>=4.60.0
pandas>=1.3.0
matplotlib>=3.3.0